requirements.txt
flask
opencv-python
ultralytics
gunicorn
numpy
pillow